no need for stand-alone j2k::TileWriter
