+++ modules/c++/nitf-c++.vcxproj.filters
@@ -211,9 +211,6 @@
     <ClCompile Include="nitf\source\J2KEncoder.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="nitf\source\J2KTileWriter.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
     <ClCompile Include="nitf\source\J2KCompressor.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
@@ -456,9 +453,6 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="nitf\include\nitf\J2KEncoder.hpp">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="nitf\include\nitf\J2KTileWriter.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="nitf\include\nitf\J2KCompressor.hpp">
