+++ modules/c++/nitf/CMakeLists.txt
@@ -54,7 +54,6 @@ coda_add_module(
         source/J2KImage.cpp
         source/J2KReader.cpp
         source/J2KStream.cpp
-        source/J2KTileWriter.cpp
         source/J2KWriter.cpp
         source/LabelSegment.cpp
         source/LabelSubheader.cpp
